[engraving] remove Qt dependencies

+++ src/importexport/midi/internal/midiimport/importmidi_tuplet_filter.cpp
@@ -442,7 +442,7 @@ bool validateSelectedTuplets(Iter beginIt,
                              const std::vector<TupletInfo>& tuplets)
 {
     // <chord address, used voices>
-    std::map<std::pair<const ReducedFraction, MidiChord>*, int> usedChords;
+    std::map<std::pair<const ReducedFraction, MidiChord>*, size_t> usedChords;
     for (auto indexIt = beginIt; indexIt != endIt; ++indexIt) {
         const auto& tuplet = tuplets[*indexIt];
         const auto& chords = tuplet.chords;
